projects/capstone/starter/requirements.txt
aniso8601==6.0.0
astroid==2.2.5
Authlib==0.15.2
Click==7.0
cryptography
ecdsa==0.13.2
Flask-Authlib-Client==0.0.1
Flask-Cors==3.0.8
Flask-Login
Flask-OAuthlib==0.9.6
Flask-RESTful==0.3.7
Flask-Session
Flask-SQLAlchemy==2.4.0
Flask==1.0.3
future==0.17.1
isort==4.3.18
itsdangerous==1.1.0
Jinja2==2.10.1
jose
jwt
lazy-object-proxy==1.4.0
MarkupSafe==1.1.1
mccabe==0.6.1
psycopg2-binary==2.8.6
pyasn1
pycparser
pycryptodome==3.9.7
PyJWT
pylint==2.3.1
python-dateutil
python-dotenv
python-jose
pytz==2020.1
requests
rsa
six
SQLAlchemy
typed-ast==1.4.1
Werkzeug
wrapt==1.11.1
gunicorn